change float column decimal points from (10,0) to (10,4)

+++ Modules/Delegate/database/migrations/2024_05_08_131905_CreateDelegateTable.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->tinyInteger('type')->nullable()->default(1);
             $table->tinyInteger('commission_type')->nullable()->default(1);
-            $table->unsignedBigInteger('commission')->nullable()->default(0);
+            $table->float('commission', 10, 4)->nullable()->default(0);
 
             $table->text('remarks')->nullable();
             $table->boolean('is_active')->nullable()->default(0);
